python-doer/requirements.txt
ConsoleMenu==1.0.1
attrs
click
